AAAAAAAAAASAASFADXCDSCACXCXZ

+++ models.py
@@ -2,6 +2,7 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
 from datetime import datetime
+from sqlalchemy import Numeric
 
 db = SQLAlchemy()
 
@@ -19,7 +20,7 @@ class Stock(db.Model):
     )
 
     value = db.Column(
-        db.BigInteger,
+        db.Numeric,
         default=0,
         nullable=False,
     )
